+++ packages/expo-modules-autolinking/src/reactNativeConfig/reactNativeConfig.types.ts
@@ -36,6 +36,8 @@ export interface RNConfigDependency {
   platforms: {
     android?: RNConfigDependencyAndroid;
     ios?: RNConfigDependencyIos;
+    tvos?: RNConfigDependencyIos;
+    macos?: RNConfigDependencyIos;
     web?: RNConfigDependencyWeb;
   };
 }
@@ -47,6 +49,12 @@ export interface RNConfigResult {
   dependencies: Record<string, RNConfigDependency>;
   project: {
     ios?: {
+      sourceDir: string;
+    };
+    tvos?: {
+      sourceDir: string;
+    };
+    macos?: {
       sourceDir: string;
     };
   };
@@ -61,6 +69,8 @@ export interface RNConfigReactNativePlatformsConfig {
   platforms?: {
     android?: RNConfigReactNativePlatformsConfigAndroid;
     ios?: RNConfigReactNativePlatformsConfigIos;
+    tvos?: RNConfigReactNativePlatformsConfigIos;
+    macos?: RNConfigReactNativePlatformsConfigIos;
     web?: RNConfigReactNativePlatformsConfigWeb;
   };
 }
@@ -93,6 +103,12 @@ export interface RNConfigReactNativeAppProjectConfig {
     packageName: string;
   };
   ios?: {
+    sourceDir: string;
+  };
+  tvos?: {
+    sourceDir: string;
+  };
+  macos?: {
     sourceDir: string;
   };
 }